refactor: 불필요한 파일 제거
