--- src/main/java/woowacourse/shoppingcart/dto/OrderRequest.java
@@ -1,24 +1,0 @@
-package woowacourse.shoppingcart.dto;
-
-import javax.validation.constraints.Min;
-import javax.validation.constraints.NotNull;
-
-public class OrderRequest {
-    @NotNull
-    private final Long cartId;
-    @Min(0)
-    private final int quantity;
-
-    public OrderRequest(final Long cartId, final int quantity) {
-        this.cartId = cartId;
-        this.quantity = quantity;
-    }
-
-    public Long getCartId() {
-        return cartId;
-    }
-
-    public int getQuantity() {
-        return quantity;
-    }
-}
