--- src/main/java/woowacourse/shoppingcart/dto/Request.java
@@ -1,9 +1,0 @@
-package woowacourse.shoppingcart.dto;
-
-public interface Request {
-    interface id {
-    }
-
-    interface allProperties {
-    }
-}
